requirements.txt
Langchain~=1.0.5
LangGraph~=1.0.3
Fastapi
torch
sentence-transformers
numpy
chromadb~=1.3.4
openpyxl
psycopg[binary]~=3.2.12
psycopg_pool
python-dotenv
psycopg2-binary
pydantic~=2.12.4
langchain-google-genai
requests~=2.32.5
langgraph-checkpoint-postgres